Split out {Monster,}FleetPlan into FleetPlan.{h,cpp}
