+++ parse/MonsterFleetPlansParser.cpp
@@ -6,7 +6,7 @@
 
 #include "../universe/Condition.h"
 #include "../universe/Universe.h"
-#include "../universe/UniverseGenerator.h"
+#include "../universe/FleetPlan.h"
 #include "../util/Directories.h"
 
 #include <boost/spirit/include/phoenix.hpp>
