+++ server/ServerWrapper.cpp
@@ -10,6 +10,7 @@
 #include "../universe/Planet.h"
 #include "../universe/Building.h"
 #include "../universe/Fleet.h"
+#include "../universe/FleetPlan.h"
 #include "../universe/Ship.h"
 #include "../universe/ShipDesign.h"
 #include "../universe/Field.h"
