+++ universe/Universe.cpp
@@ -14,12 +14,12 @@
 #include "Building.h"
 #include "Effect.h"
 #include "Fleet.h"
+#include "FleetPlan.h"
 #include "Planet.h"
 #include "Ship.h"
 #include "System.h"
 #include "Field.h"
 #include "UniverseObject.h"
-#include "UniverseGenerator.h"
 #include "UnlockableItem.h"
 #include "Predicates.h"
 #include "ShipPart.h"
