+++ universe/UniverseGenerator.h
@@ -3,65 +3,7 @@
 
 #include "Universe.h"
 
-
 struct PlayerSetupData;
-
-/** A combination of names of ShipDesign that can be put together to make a
- * fleet of ships, and a name for such a fleet, loaded from
- * default/scripting/starting_unlocks/fleets.inf
- * ShipDesign names refer to designs listed in default/scripting/ship_designs.
- * Useful for saving or specifying prearranged combinations of prearranged
- * ShipDesigns to automatically put together, such as during universe creation.*/
-class FleetPlan {
-public:
-    FleetPlan(const std::string& fleet_name, const std::vector<std::string>& ship_design_names,
-              bool lookup_name_userstring = false) :
-        m_name(fleet_name),
-        m_ship_designs(ship_design_names),
-        m_name_in_stringtable(lookup_name_userstring)
-    {}
-    FleetPlan() :
-        m_name(""),
-        m_ship_designs(),
-        m_name_in_stringtable(false)
-    {}
-    virtual ~FleetPlan() {};
-    const std::string&              Name() const;
-    const std::vector<std::string>& ShipDesigns() const { return m_ship_designs; }
-
-protected:
-    std::string                     m_name;
-    std::vector<std::string>        m_ship_designs;
-    bool                            m_name_in_stringtable;
-};
-
-/** The combination of a FleetPlan and spawning instructions for start of game
-  * monsters. */
-class FO_COMMON_API MonsterFleetPlan : public FleetPlan {
-public:
-    MonsterFleetPlan(const std::string& fleet_name, const std::vector<std::string>& ship_design_names,
-                     double spawn_rate = 1.0, int spawn_limit = 9999,
-                     std::unique_ptr<Condition::Condition>&& location = nullptr,
-                     bool lookup_name_userstring = false) :
-        FleetPlan(fleet_name, ship_design_names, lookup_name_userstring),
-        m_spawn_rate(spawn_rate),
-        m_spawn_limit(spawn_limit),
-        m_location(std::move(location))
-    {}
-    MonsterFleetPlan() :
-        FleetPlan()
-    {}
-    virtual ~MonsterFleetPlan()
-    {}
-    double                          SpawnRate() const   { return m_spawn_rate; }
-    int                             SpawnLimit() const  { return m_spawn_limit; }
-    const Condition::Condition* Location() const    { return m_location.get(); }
-protected:
-    double                          m_spawn_rate = 1.0;
-    int                             m_spawn_limit = 9999;
-    // Use shared_ptr insead of unique_ptr because boost::python requires a deleter
-    const std::shared_ptr<Condition::Condition> m_location;
-};
 
 /** Set active meter current values equal to target/max meter current
  * values.  Useful when creating new object after applying effects. */
